Modify the name of engine about spark streaming
