+++ waterdrop-core/src/main/resources/META-INF/services/io.github.interestinglab.waterdrop.apis.BaseStreamingInput
@@ -1,6 +1,6 @@
-io.github.interestinglab.waterdrop.input.streaming.FakeStream
-io.github.interestinglab.waterdrop.input.streaming.FileStream
-io.github.interestinglab.waterdrop.input.streaming.HdfsStream
-io.github.interestinglab.waterdrop.input.streaming.KafkaStream
-io.github.interestinglab.waterdrop.input.streaming.S3Stream
-io.github.interestinglab.waterdrop.input.streaming.SocketStream
+io.github.interestinglab.waterdrop.input.sparkstreaming.FakeStream
+io.github.interestinglab.waterdrop.input.sparkstreaming.FileStream
+io.github.interestinglab.waterdrop.input.sparkstreaming.HdfsStream
+io.github.interestinglab.waterdrop.input.sparkstreaming.KafkaStream
+io.github.interestinglab.waterdrop.input.sparkstreaming.S3Stream
+io.github.interestinglab.waterdrop.input.sparkstreaming.SocketStream
