+++ waterdrop-core/src/main/scala/io/github/interestinglab/waterdrop/input/sparkstreaming/FakeStream.scala
@@ -1,4 +1,4 @@
-package io.github.interestinglab.waterdrop.input.streaming
+package io.github.interestinglab.waterdrop.input.sparkstreaming
 
 import java.security.SecureRandom
 import java.util
