+++ waterdrop-core/src/main/scala/io/github/interestinglab/waterdrop/input/sparkstreaming/S3Stream.scala
@@ -1,4 +1,4 @@
-package io.github.interestinglab.waterdrop.input.streaming
+package io.github.interestinglab.waterdrop.input.sparkstreaming
 
 import com.typesafe.config.{Config, ConfigFactory}
 import io.github.interestinglab.waterdrop.apis.BaseStreamingInput
@@ -8,9 +8,7 @@ import org.apache.spark.sql.{Dataset, Row, RowFactory, SparkSession}
 import org.apache.spark.streaming.StreamingContext
 import org.apache.spark.streaming.dstream.DStream
 
-import scala.collection.JavaConversions._
-
-class SocketStream extends BaseStreamingInput[String] {
+class S3Stream extends BaseStreamingInput[String] {
 
   var config: Config = ConfigFactory.empty()
 
@@ -28,22 +26,29 @@ class SocketStream extends BaseStreamingInput[String] {
     this.config
   }
 
-  override def checkConfig(): (Boolean, String) = (true, "")
+  override def checkConfig(): (Boolean, String) = {
+    config.hasPath("path") match {
+      case true => {
+        val allowedURISchema = List("s3://", "s3a://", "s3n://")
+        val dir = config.getString("path")
+        val unSupportedSchema = allowedURISchema.forall(schema => {
+          // there are 3 "/" in dir, first 2 are from URI Schema, and the last is from path
+          !dir.startsWith(schema + "/")
+        })
 
-  override def prepare(spark: SparkSession): Unit = {
-    super.prepare(spark)
-
-    val defaultConfig = ConfigFactory.parseMap(
-      Map(
-        "host" -> "localhost",
-        "port" -> 9999
-      ))
-    config = config.withFallback(defaultConfig)
+        unSupportedSchema match {
+          case true =>
+            (false, "unsupported schema, please set the following allowed schemas: " + allowedURISchema.mkString(", "))
+          case false => (true, "")
+        }
+      }
+      case false => (false, "please specify [path] as non-empty string")
+    }
   }
 
   override def getDStream(ssc: StreamingContext): DStream[String] = {
 
-    ssc.socketTextStream(config.getString("host"), config.getInt("port"))
+    ssc.textFileStream(config.getString("path"))
   }
 
   override def rdd2dataset(spark: SparkSession, rdd: RDD[String]): Dataset[Row] = {
